[refactor] #334 온보딩 ViewHolder 분리
