+++ presentation/src/main/java/kr/techit/lion/presentation/splash/adapter/OnBoardingImageVPAdapter.kt
@@ -1,34 +1,22 @@
 package kr.techit.lion.presentation.splash.adapter
 
-import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import kr.techit.lion.presentation.databinding.ItemOnboardingVpBinding
 import kr.techit.lion.presentation.onboarding.OnBoardingPage
+import kr.techit.lion.presentation.splash.adapter.OnBoardingViewHolder.Companion.OnBoardingViewHolder
 
 class OnBoardingImageVPAdapter(
     private val pages: List<OnBoardingPage>,
-) :  RecyclerView.Adapter<OnBoardingImageVPAdapter.ImageViewHolder>() {
+) : RecyclerView.Adapter<OnBoardingViewHolder>() {
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ImageViewHolder {
-        val binding : ItemOnboardingVpBinding = ItemOnboardingVpBinding.inflate(
-            LayoutInflater.from(parent.context), parent, false)
-
-        return ImageViewHolder(binding)
-    }
+    override fun onCreateViewHolder(
+        parent: ViewGroup,
+        viewType: Int,
+    ): OnBoardingViewHolder = OnBoardingViewHolder(parent)
 
     override fun getItemCount(): Int = pages.size
 
-    override fun onBindViewHolder(holder: ImageViewHolder, position: Int) {
+    override fun onBindViewHolder(holder: OnBoardingViewHolder, position: Int) {
         holder.bind(pages[position])
-    }
-
-    class ImageViewHolder(private val binding: ItemOnboardingVpBinding) : RecyclerView.ViewHolder(binding.root) {
-        fun bind(page: OnBoardingPage) {
-            binding.itemOnboardingIv.setImageDrawable(page.image)
-            binding.itemOnboardingTv1.text = page.title
-            binding.itemOnboardingTv2.text = page.description
-            binding.itemOnboardingTv3.text = page.extra
-        }
     }
 }
